Update 1
Second upload of the DailyMath program. The functions themselves now pop up in a window after the user picks what they want to do. Currently, the cancel option causes the program to crash, but the program functions as intended currently. It runs on a loop and allows the user to complete 10 problems before exiting.